Create test to check the CorrelationID bug is still present.
The CorrelationIdMiddleware class has a comment that 'in .Net core 2.1+ a log property called 'CorrelationID' is replaced with null'.

An integration test has been created to test for this issue.  It uses the Serilog Correlation Log Enricher, which adds a property called CorrelationID.  I presume that this was the original occurrence of the bug, given that the Serilog LogContext is used in the middleware.

Other integration tests have been included to test the other features of the middleware for completeness.